// src/types/ProfileData.ts
export interface ProfileData {
  id: number;
  name: string;
  type: string;
  manager: string;
  department: string;
  center: string;
  calendarUntil: string;
  lastSync: string;
  currentStatus: string;
  currentFreeCapacity: string;
  timeZone: string;
  resourceTime: string;
  startLocation: string;
  currentLocation: string;
  endLocation: string;
  phone: string;
  email: string;
  mobilePhone: string;
  fullDayAvailable: string;
  skills: string[];
}